unificando controllers restaurante
